requirements.txt
django==5.1.6
psycopg==3.2.4
psycopg-binary==3.2.4
djangorestframework==3.15.2
dj-database-url==2.3.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
requests==2.32.3
